requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
openai
requests
beautifulsoup4
nltk
pydantic
groq